routing: add routing to all pages and use Link tag for better user experience
