+++ frontend/src/pages/HomePage.tsx
@@ -1,7 +1,10 @@
 import { BackgroundBeams } from "../components/ui/background-beams";
 import Button from "../components/Button";
+import { useNavigate } from "react-router-dom";
 
 const HomePage: React.FC = () => {
+  const navigate = useNavigate();
+
   return (
     <div className="h-screen w-full bg-neutral-950 relative flex flex-col items-center justify-center antialiased">
       <div className="max-w-2xl mx-auto p-4 relative z-10 text-center">
@@ -15,7 +18,7 @@ const HomePage: React.FC = () => {
         </p>
         <div className="flex justify-center mt-6">
           {" "}
-          <Button label="Get Started Now" />
+          <Button label="Get Started Now" onClick={() => navigate("/signup")} />
         </div>
       </div>
       <BackgroundBeams className="absolute inset-0 z-0" />
