+++ frontend/src/pages/auth/ForgotPasswordPage.tsx
@@ -1,60 +1,9 @@
-// import { useState } from "react";
-// import Button from "../../components/Button";
-// import FormWrapper from "../../components/FormWrapper";
-// import Input from "../../components/Input";
-// import TextButtonWithIcon from "../../components/TextButtonWithIcon";
-
-// const ForgotPasswordPage: React.FC = () => {
-//   const [formData, setFormData] = useState({
-//     email: "",
-//     termsAccepted: false,
-//   });
-
-//   const handleSubmit = (e: React.FormEvent) => {
-//     e.preventDefault();
-//   };
-//   const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-//     const { name, value } = e.target;
-//     setFormData((prevData) => ({
-//       ...prevData,
-//       [name]: value,
-//     }));
-//   };
-//   const handleBackToLogin = () => {};
-
-//   return (
-//     <FormWrapper title="Forgot your password?">
-//       <form onSubmit={handleSubmit}>
-//         <Input
-//           label="Your Email"
-//           type="email"
-//           name="email"
-//           placeholder="email@gmail.com"
-//           required
-//           fullWidth
-//           value={formData.email}
-//           onChange={handleInputChange}
-//         />
-
-//         <div className="flex justify-between items-center w-full mt-4">
-//           <TextButtonWithIcon
-//             onClick={handleBackToLogin}
-//             label="Back to Login"
-//           />
-//           <Button type="submit" label="Reset Password" variant="filled" />
-//         </div>
-//       </form>
-//     </FormWrapper>
-//   );
-// };
-
-// export default ForgotPasswordPage;
-
-import { FieldValues, useForm } from "react-hook-form";
-import { z } from "zod";
 import { zodResolver } from "@hookform/resolvers/zod";
-import Button from "../../components/Button";
+import { FieldValues, useForm } from "react-hook-form";
 import { FiArrowLeft } from "react-icons/fi";
+import { Link, useNavigate } from "react-router-dom";
+import { z } from "zod";
+import Button from "../../components/Button";
 
 const schema = z.object({
   email: z
@@ -72,11 +21,12 @@ const ForgotPasswordPage: React.FC = () => {
     handleSubmit,
   } = useForm<FormData>({ resolver: zodResolver(schema) });
 
-  const onSubmit = (data: FieldValues) => {
-    console.log("Form Data:", data);
-  };
+  const navigate = useNavigate();
 
-  const handleBackToLogin = () => {};
+  const onSubmit = (data: FieldValues) => {
+    console.log(data);
+    navigate("/otpVerify");
+  };
 
   return (
     <section className="bg-gray-50 dark:bg-gray-900">
@@ -112,13 +62,13 @@ const ForgotPasswordPage: React.FC = () => {
               </div>
 
               <div className="flex justify-between items-center">
-                <button
-                  onClick={handleBackToLogin}
+                <Link
+                  to="/login"
                   className="flex items-center text-gray-500 hover:underline focus:outline-none"
                 >
                   <FiArrowLeft className="mr-2" />
                   Back to Login
-                </button>
+                </Link>
 
                 <Button
                   type="submit"
